requirements.txt
Django==1.4.16
MySQL-python==1.2.5
Pillow==2.3.0
South==1.0.1
argparse==1.2.1
distribute==0.6.24
django-extensions==0.8
django-mptt==0.5.2
django-registration==1.0
django-robots==1.0
django-simple-captcha==0.4.3
django-tagging==0.3.1
gunicorn==18.0
psycopg2==2.5.4
sorl-thumbnail==12.3
wsgiref==0.1.2
